Improve algorithm profitability and portfolio size adaptability
Key changes across 4 files:

execution/alpaca_adapter.py:
- Replace int() truncation with round(..., 6) for all position quantities
  so Alpaca fractional share support is fully used; critical for expensive
  stocks (NVDA/MSFT) in small portfolios

council/portfolio.py:
- Add _get_portfolio_tier(): adaptive constraints by portfolio value
    <$5k   → top-3 signals, 45% max, $50 min trade
    $5-25k → top-5 signals, 25% max, $25 min trade
    $25-100k → top-10 signals, 15% max, $10 min trade
    >$100k → full universe, original limits
- Set commission_bps=0.0 (Alpaca charges 0 on equities)
- Add signal gate: zero out signals where |z-score| < 0.20 (noise filter)
- Add rebalance cool-down: small portfolios skip if rebalanced <5d ago
- Add apply_drawdown_scale(): circuit breaker that scales positions down
  linearly when 5d portfolio return exceeds -7% threshold
- compute_orders() uses tier-based min_trade_usd instead of hardcoded $1

council/aggregator.py:
- Fix weight projection: replace simple clip-renorm (which violates
  max_weight after renormalization) with a correct fixed/free redistribution
  algorithm that projects onto the [min_w, max_w] simplex exactly
- Use soft floor max(0.1, sharpe) instead of max(0.0, sharpe) so
  temporarily underperforming models retain a minimum contribution
- Replace rolling mean IC-Sharpe with EWM (halflife=20d) for faster
  adaptation to regime changes

council/conformal.py:
- Raise filter_low_confidence threshold from 80th to 90th percentile,
  preserving more investable universe (drop 10% instead of 20%)

+++ execution/alpaca_adapter.py
@@ -270,7 +270,7 @@ class AlpacaLiveNode:
     def _submit_crypto_order(
         self,
         symbol: str,
-        qty: int,
+        qty: float,
         side: str,
         order_type: str = "market",
         time_in_force: str = "day",
@@ -326,7 +326,7 @@ class AlpacaLiveNode:
     def _submit_equity_order(
         self,
         symbol: str,
-        qty: int,
+        qty: float,
         side: str,
         order_type: str = "market",
         time_in_force: str = "day",
@@ -374,7 +374,7 @@ class AlpacaLiveNode:
     def submit_order(
         self,
         symbol: str,
-        qty: int,
+        qty: float,
         side: str,
         order_type: str = "market",
         time_in_force: str = "day",
